refactor: 카테고리의 상세내역 화면을 구현했습니다.
화면에서 아이템 카드를 롱클릭하면 삭제 다이얼로그 뜨도록 하였고, 일반클릭을 하면 수정 다이얼로그가 뜨도록 하였습니다.

해당 화면에서 용돈일기 다이얼로그가 뜨면 타입 수정은 안되도록 지정했습니다.

날짜도 년도랑 월도 변경 못하게 하려고 했지만 유저의 편의성을 위해 하지 않았습니다.

해당 화면에서의 용돈일기 다이얼로그에서는 카테고리 변경을 지원하지 않습니다.

리포지토리와 유스케이스 id 타입을 Long으로 수정했습니다.

루트에서 속성을 지정해줬습니다.

기존 용돈일기 다이얼로그의 isReplace 매개변수를 isTypeLock 으로 이름 변경했습니다.

그 외에 의존성 수동 주입과 뷰모델 생성 코드를 작성했습니다.

+++ common/src/main/java/com/paykidscompose/common/repositories/ExpenseAllowanceRepository.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.Flow
 import java.time.LocalDate
 
 interface ExpenseAllowanceRepository {
-    suspend fun deleteExpense(id: Int): DataResourceResult<Boolean>
+    suspend fun deleteExpense(id: Long): DataResourceResult<Boolean>
     suspend fun saveExpense(request: AllowanceChartModel): DataResourceResult<Boolean>
     suspend fun replaceExpense(request: AllowanceChartModel): DataResourceResult<Boolean>
     suspend fun getExpenseMonthTotalAmount(year: Int, month: Int): DataResourceResult<Int>
